Move exporter's e2e test to linera-service crate
